Add corner detection and edge detection with double ray hit
The rays now identify when a corner is between them and add vertices to the corner. The corner detection only works with sharp edges.
In cases where the latest 2 rays get a hit but the latest one hit a collider far behind the initial one, an edge is detected and vertices are added on the edge.

WIP #103